Refactor deterministic and benchmark logic
